requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
sqlalchemy==2.0.25

# AI/ML Models
torch==2.1.0
diffusers==0.24.0
transformers==4.35.0
scipy==1.11.0
numpy==1.24.0

# 3D Models
trimesh==3.17.0
numpy-stl==3.0.1
